# openapi.yaml
openapi: 3.0.3
info:
  title: Nutri Thata API
  description: API para plataforma de educação nutricional com sistema de módulos, convites e upload de mídia
  version: 1.0.0
  contact:
    name: API Support
    email: support@nutrithata.com
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://api.nutrithata.com
    description: Produção
security:
  - BearerAuth: []
paths:
...
  /modules:
    get:
      tags:
        - Módulos
      summary: Listar módulos
      description: Lista módulos educacionais com filtros e paginação
      parameters:
        - name: category
          in: query
          description: Filtrar por categoria
          schema:
            type: string
            example: "Fundamentos"
        - name: page
          in: query
          description: Número da página (base 0)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Tamanho da página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Campo de ordenação
          schema:
            type: string
            enum: ["title", "createdAt", "category"]
            default: "createdAt"
        - name: direction
          in: query
          description: Direção da ordenação
          schema:
            type: string
            enum: ["asc", "desc"]
            default: "desc"
      responses:
        '200':
          description: Lista de módulos
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      modules:
                        type: array
                        items:
                          $ref: '#/components/schemas/ModuleSummary'
                      totalElements:
                        type: integer
                        example: 50
                      totalPages:
                        type: integer
                        example: 5
                      currentPage:
                        type: integer
                        example: 0
                      pageSize:
                        type: integer
                        example: 20
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Módulos
      summary: Criar novo módulo
      description: Cria módulo educacional (PROFESSIONAL ou ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleCreateRequest'
      responses:
        '201':
          description: Módulo criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Módulo criado com sucesso"
                  data:
                    $ref: '#/components/schemas/Module'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão (apenas PROFESSIONAL ou ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /modules/{id}:
    get:
      tags:
        - Módulos
      summary: Obter módulo por ID
      description: Retorna módulo completo com conteúdo
      parameters:
        - name: id
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Módulo encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    $ref: '#/components/schemas/Module'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Módulo não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Módulos
      summary: Atualizar módulo
      description: Atualiza módulo (criador ou ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleUpdateRequest'
      responses:
        '200':
          description: Módulo atualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Módulo atualizado com sucesso"
                  data:
                    $ref: '#/components/schemas/Module'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão (apenas criador ou ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Módulo não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Módulos
      summary: Excluir módulo
      description: Remove módulo (criador ou ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Módulo excluído com sucesso
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão (apenas criador ou ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Módulo não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== UPLOAD DE MÍDIA ==========
  /media/upload:
    post:
      tags:
        - Mídia
      summary: Upload de arquivo
      description: Faz upload de arquivo local (imagem, vídeo ou áudio)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Arquivo para upload
                type:
                  type: string
                  enum: ["IMAGE", "VIDEO", "AUDIO"]
                  description: Tipo de mídia (auto-detectado se não informado)
      responses:
        '201':
          description: Upload realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Upload concluído"
                  data:
                    $ref: '#/components/schemas/MediaAsset'
        '400':
          description: Tipo de arquivo não suportado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Arquivo muito grande
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /media/link:
    post:
      tags:
        - Mídia
      summary: Vincular URL externa
      description: Salva referência para mídia em URL externa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - type
              properties:
                url:
                  type: string
                  format: uri
                  example: "https://youtube.com/embed/dQw4w9WgXcQ"
                type:
                  type: string
                  enum: ["IMAGE", "VIDEO", "AUDIO"]
                  example: "VIDEO"
      responses:
        '201':
          description: URL validada e salva
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "URL validada e salva"
                  data:
                    $ref: '#/components/schemas/MediaAsset'
        '400':
          description: URL inválida ou inacessível
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /media/my:
    get:
      tags:
        - Mídia
      summary: Listar mídias do usuário
      description: Lista arquivos de mídia do usuário autenticado
      parameters:
        - name: type
          in: query
          description: Filtrar por tipo de mídia
          schema:
            type: string
            enum: ["IMAGE", "VIDEO", "AUDIO"]
        - name: page
          in: query
          description: Número da página (base 0)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Tamanho da página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Lista de mídias
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      media:
                        type: array
                        items:
                          $ref: '#/components/schemas/MediaAsset'
                      totalElements:
                        type: integer
                        example: 25
                      totalPages:
                        type: integer
                        example: 3
                      currentPage:
                        type: integer
                        example: 0
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /media/{id}:
    delete:
      tags:
        - Mídia
      summary: Excluir mídia
      description: Remove arquivo de mídia (dono ou ADMIN)
      parameters:
        - name: id
          in: path
          required: true
          description: ID da mídia
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Mídia removida com sucesso
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Sem permissão (apenas dono ou ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Mídia não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== CONVITES ==========
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        faultstring:
          type: string
          example: "Credenciais inválidas"
        stacktrace:
          type: string
          description: "Disponível apenas em ambiente de desenvolvimento"
          example: "com.nutrithata.exception.AuthenticationException: Invalid credentials..."

    # ========== MODELOS DE USUÁRIO ==========
...
    ModuleSummary:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "Introdução à Alimentação Saudável"
        description:
          type: string
          example: "Fundamentos básicos para uma alimentação equilibrada e nutritiva"
        coverImage:
          type: string
          example: "https://api.nutrithata.com/static/covers/intro-alimentacao.jpg"
        category:
          type: string
          example: "Fundamentos"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        contentCount:
          type: integer
          example: 3
        createdBy:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: "Dra. Ana Paula Silva"
    Module:
      allOf:
        - $ref: '#/components/schemas/ModuleSummary'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ContentBlock'
            updatedAt:
              type: string
              format: date-time
              example: "2024-01-20T15:45:00Z"
    ContentBlock:
      type: object
      properties:
        id:
          type: string
          example: "1-1"
        type:
          type: string
          enum: ["text", "video", "audio"]
          example: "text"
        content:
          type: string
          example: "A alimentação saudável é a base para uma vida plena e cheia de energia..."
        order:
          type: integer
          example: 1
    ModuleCreateRequest:
      type: object
      required:
        - title
        - description
        - category
        - content
      properties:
        title:
          type: string
          example: "Nutrição na Terceira Idade"
        description:
          type: string
          example: "Cuidados nutricionais específicos para idosos"
        coverImage:
          type: string
          example: "https://api.nutrithata.com/static/covers/terceira-idade.jpg"
        category:
          type: string
          example: "Grupos Especiais"
        content:
          type: array
          items:
            type: object
            required:
              - type
              - content
              - order
            properties:
              type:
                type: string
                enum: ["text", "video", "audio"]
              content:
                type: string
              order:
                type: integer
    ModuleUpdateRequest:
      type: object
      properties:
        title:
          type: string
          example: "Nutrição na Terceira Idade - Atualizado"
        description:
          type: string
          example: "Cuidados nutricionais específicos para idosos com foco em prevenção"
        coverImage:
          type: string
          example: "https://api.nutrithata.com/static/covers/terceira-idade-v2.jpg"
        category:
          type: string
          example: "Grupos Especiais"
        content:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: "ID do bloco existente (para atualização) ou null (para criação)"
              type:
                type: string
                enum: ["text", "video", "audio"]
              content:
                type: string
              order:
                type: integer

    # ========== MODELOS DE MÍDIA ==========
    MediaAsset:
      type: object
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        ownerUserId:
          type: integer
          format: int64
          example: 1
        type:
          type: string
          enum: ["IMAGE", "VIDEO", "AUDIO"]
          example: "IMAGE"
        storage:
          type: string
          enum: ["LOCAL", "EXTERNAL_URL"]
          example: "LOCAL"
        filePath:
          type: string
          nullable: true
          example: "uploads/2024/01/abc123-image.jpg"
        publicUrl:
          type: string
          nullable: true
          example: "https://api.nutrithata.com/static/uploads/2024/01/abc123-image.jpg"
        externalUrl:
          type: string
          nullable: true
          example: "https://youtube.com/embed/dQw4w9WgXcQ"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"

    # ========== MODELOS DE CONVITE ==========
...